Flatten preprint form structure
