--- app/templates/components/preprint-form-basics.hbs
@@ -1,20 +1,0 @@
-<div class="col-md-6">
-    <label>
-        <span class="required">Title:</span>
-        {{input class="form-control" value=title required=true}}
-    </label>
-    <label>
-        DOI:
-        {{input class="form-control" value=doi}}
-    </label>
-    <label>
-        Tags:
-        {{input class="form-control" value=tags}}
-    </label>
-</div>
-<div class="col-md-6">
-    <label>
-        <span class="required">Abstract:</span>
-        {{textarea class="form-control" value=abstract required=true}}
-    </label>
-</div>
