+++ app/templates/components/preprint-form-body.hbs
@@ -1,0 +1,1 @@
+{{yield}}
