--- app/templates/components/preprint-form-upload.hbs
@@ -1,39 +1,0 @@
-{{#liquid-bind state class='translate' as |currentState|}}
-    <div class={{currentState}}>
-        {{#if (eq currentState _State.START)}}
-            <button class="col-md-4"{{action 'changeState' _State.NEW}}>Upload a New File</button>
-            <button class="col-md-4"{{action 'changeState' _State.EXISTING}}>Select a file from OSF</button>
-
-        {{else if (eq currentState _State.NEW)}}
-            {{dropzone-widget buildUrl="/"}}
-            {{#liquid-bind uploadState use='crossFade' as |currentUploadState|}}
-                {{#if (eq currentUploadState _State.START)}}
-                    <button class="btn btn-primary" {{action 'changeUploadState' _State.NEW}}>Create new project</button>
-                    <button class="btn btn-primary" {{action 'changeUploadState' _State.EXISTING}}>Save to existing OSF project</button>
-                    <button class="btn btn-default" {{action 'changeState' _State.START}}>Back</button>
-
-                {{else if (eq currentUploadState _State.NEW)}}
-                    <h1>Creating new project</h1>
-                    <button class="btn btn-default" {{action 'changeUploadState' _State.START}}>Back</button>
-                    <button class="btn btn-default" {{action next}}>Next</button>
-
-                {{else if (eq currentUploadState _State.EXISTING)}}
-                    {{preprint-form-project-select userNodes=userNodes}}
-                    <button class="btn btn-default" {{action 'changeUploadState' _State.START}}>Back</button>
-                    <button class="btn btn-primary" {{action next}}>Next</button>
-
-                {{else}}
-                    {{log 'preprint-form-upload: bad uploadState'}}
-                {{/if}}
-            {{/liquid-bind}}
-
-        {{else if (eq currentState _State.EXISTING)}}
-            {{preprint-form-project-select userNodes=userNodes fileSelect=true}}
-            <button class="btn btn-default" {{action 'changeState' _State.START}}>Back</button>
-            <button class="btn btn-primary" {{action next}}>Next</button>
-
-        {{else}}
-            {{log 'preprint-form-upload: bad state'}}
-        {{/if}}
-    </div>
-{{/liquid-bind}}
