--- app/templates/components/preprint-form.hbs
@@ -1,37 +1,0 @@
-{{#with _names.[0] as |name|}}
-    {{#preprint-form-section id='preprint-form-upload' open=true name=name}}
-        {{preprint-form-header name=name valid=(get valid name)}}
-        {{preprint-form-upload verify=(action 'verify' name) next=(action 'next' name)}}
-    {{/preprint-form-section}}
-{{/with}}
-
-{{#with _names.[1] as |name|}}
-    {{#preprint-form-section id='preprint-form-basics' name=name}}
-        {{preprint-form-header name=name valid=(get valid name)}}
-        {{preprint-form-basics verify=(action 'verify' name)}}
-    {{/preprint-form-section}}
-{{/with}}
-
-{{#with _names.[2] as |name|}}
-    {{#preprint-form-section id='preprint-form-subjects' name=name}}
-        {{preprint-form-header name=name valid=(get valid name)}}
-        {{preprint-form-subjects verify=(action 'verify' name)}}
-    {{/preprint-form-section}}
-{{/with}}
-
-{{#with _names.[3] as |name|}}
-    {{#preprint-form-section id='preprint-form-authors' name=name}}
-        {{preprint-form-header name=name valid=(get valid name)}}
-        {{preprint-form-authors verify=(action 'verify' name)}}
-    {{/preprint-form-section}}
-{{/with}}
-
-{{#with _names.[4] as |name|}}
-    {{#preprint-form-section id='preprint-form-submit' animate=true name=name}}
-        {{preprint-form-header name=name valid=(get valid name)}}
-        {{#cp-panel-body}}
-            <p>Sharing this project will make both the manuscript and the project associated with it public.</p>
-            <button class="btn btn-primary" type="submit">Share</button>
-        {{/cp-panel-body}}
-    {{/preprint-form-section}}
-{{/with}}
