+++ tests/integration/components/preprint-form-body-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('preprint-form-body', 'Integration | Component | preprint form body', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{preprint-form-body}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#preprint-form-body}}
+      template block text
+    {{/preprint-form-body}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
